+++ lib/src/view/extensions/_extensions_view.dart
@@ -1,0 +1,31 @@
+///
+/// Copyright (C) 2022 Andrious Solutions
+///
+/// Licensed under the Apache License, Version 2.0 (the "License");
+/// you may not use this file except in compliance with the License.
+/// You may obtain a copy of the License at
+///
+///    http://www.apache.org/licenses/LICENSE-2.0
+///
+/// Unless required by applicable law or agreed to in writing, software
+/// distributed under the License is distributed on an "AS IS" BASIS,
+/// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+/// See the License for the specific language governing permissions and
+/// limitations under the License.
+///
+///          Created  15 May 2022
+///
+
+export 'package:mvc_application/src/view/extensions/context_extensions.dart';
+
+export 'package:mvc_application/src/view/extensions/double_extensions.dart';
+
+export 'package:mvc_application/src/view/extensions/duration_extensions.dart';
+
+export 'package:mvc_application/src/view/extensions/dynamic_extensions.dart';
+
+export 'package:mvc_application/src/view/extensions/num_extensions.dart';
+
+export 'package:mvc_application/src/view/extensions/string_extensions.dart';
+
+export 'package:mvc_application/src/view/extensions/widget_extensions.dart';