+++ lib/src/view/extensions/widget_extensions.dart
@@ -1,0 +1,80 @@
+/// MIT License
+///
+/// Copyright (c) 2019 Jonny Borges
+///
+/// Permission is hereby granted, free of charge, to any person obtaining a copy
+/// of this software and associated documentation files (the "Software"), to deal
+/// in the Software without restriction, including without limitation the rights
+/// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+/// copies of the Software, and to permit persons to whom the Software is
+/// furnished to do so, subject to the following conditions:
+///
+/// The above copyright notice and this permission notice shall be included in all
+/// copies or substantial portions of the Software.
+///
+/// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+/// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+/// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+/// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+/// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+/// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+/// SOFTWARE.
+///
+/// Original source: get 4.6.1
+
+import 'package:flutter/widgets.dart';
+
+/// add Padding Property to widget
+extension WidgetPaddingX on Widget {
+  Widget paddingAll(double padding) =>
+      Padding(padding: EdgeInsets.all(padding), child: this);
+
+  Widget paddingSymmetric({double horizontal = 0.0, double vertical = 0.0}) =>
+      Padding(
+          padding:
+              EdgeInsets.symmetric(horizontal: horizontal, vertical: vertical),
+          child: this);
+
+  Widget paddingOnly({
+    double left = 0.0,
+    double top = 0.0,
+    double right = 0.0,
+    double bottom = 0.0,
+  }) =>
+      Padding(
+          padding: EdgeInsets.only(
+              top: top, left: left, right: right, bottom: bottom),
+          child: this);
+
+  Widget get paddingZero => Padding(padding: EdgeInsets.zero, child: this);
+}
+
+/// Add margin property to widget
+extension WidgetMarginX on Widget {
+  Widget marginAll(double margin) =>
+      Container(margin: EdgeInsets.all(margin), child: this);
+
+  Widget marginSymmetric({double horizontal = 0.0, double vertical = 0.0}) =>
+      Container(
+          margin:
+              EdgeInsets.symmetric(horizontal: horizontal, vertical: vertical),
+          child: this);
+
+  Widget marginOnly({
+    double left = 0.0,
+    double top = 0.0,
+    double right = 0.0,
+    double bottom = 0.0,
+  }) =>
+      Container(
+          margin: EdgeInsets.only(
+              top: top, left: left, right: right, bottom: bottom),
+          child: this);
+
+  Widget get marginZero => Container(margin: EdgeInsets.zero, child: this);
+}
+
+/// Allows you to insert widgets inside a CustomScrollView
+extension WidgetSliverBoxX on Widget {
+  Widget get sliverBox => SliverToBoxAdapter(child: this);
+}